requirements.txt
Flask~=3.1.1
flask_sqlalchemy==3.1.1
SQLAlchemy~=2.0.41
Flask-WTF==1.2.1
WTForms~=3.2.1
Flask-Bootstrap~=3.3.7.1
Bootstrap_Flask==2.2.0
Requests~=2.32.4
python-dotenv~=1.0.1